(*) Modified implosion test to match newest master branch. Implosion test is now     in app/id_generators/implosion
