+++ app/id_generators/include/implosion.h
@@ -23,12 +23,10 @@
  * @brief User define for dimension and type 
  */
 
-#ifndef _user_h_
-#define _user_h_
+#ifndef _implosion_h_
+#define _implosion_h_
 
 #define INTERNAL_ENERGY
 #define OUTPUT
-static const size_t gdimension = 2;
-using type_t = double;
 
-#endif // _user_h_
+#endif // _implosion_h_
