+++ config/project.cmake
@@ -110,10 +110,6 @@ cinch_add_application_directory("include/physics/test")
 #------------------------------------------------------------------------------#
 # Add application targets
 #------------------------------------------------------------------------------#
-cinch_add_application_directory("app/bns_3D")
-cinch_add_application_directory("app/sedov")
-cinch_add_application_directory("app/noh")
-cinch_add_application_directory("app/implosion")
 cinch_add_application_directory("app/fluid_3D")
 cinch_add_application_directory("app/fluid_2D")
 cinch_add_application_directory("app/id_generators")
