+++ data/implosion_sqn100.par
@@ -1,0 +1,24 @@
+#
+# 2-zone implosion test
+#
+# initial data
+  initial_data_prefix = "implosion_sqn100"
+  lattice_nx = 100         # particle lattice dimension
+  poly_gamma = 1.4         # polytropic index
+  rho_initial = 1.0
+  pressure_initial = 1.0e-7
+  sph_separation = 0.001   # TODO: use sph_eta instead
+  #sph_eta = 1.5
+  sedov_blast_energy = 1.0
+  sedov_blast_radius = 1.0 # in units of particle separation
+  lattice_type = 1         # 0:rectangular, 1:hcp, 2:fcc  **in 2d both hcp and fcc are triangular**
+  domain_type = 1          # 0:cube, 1:sphere, 2:full box
+
+# evolution parameters:
+  sph_kernel = "Quintic Spline"
+  initial_dt = 0.0001 # TODO: better use Courant factor X sph_separation
+  final_iteration = 50
+  out_screen_every = 1
+  out_scalar_every = 10
+  out_h5data_every = 10
+  output_h5data_prefix = "implosion_evolution"
